[Refactoring] main.cpp의 코드를 Account.cpp로 이동
